+++ src/main/java/org/ezcode/codetest/infrastructure/persistence/repository/community/vote/DiscussionVoteJpaJpaRepository.java
@@ -1,0 +1,14 @@
+package org.ezcode.codetest.infrastructure.persistence.repository.community.vote;
+
+import java.util.Optional;
+
+import org.ezcode.codetest.domain.community.model.DiscussionVote;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DiscussionVoteJpaJpaRepository extends JpaRepository<DiscussionVote, Long> {
+
+	Optional<DiscussionVote> findByVoterIdAndDiscussionId(Long voterId, Long discussionId);
+
+	boolean existsByVoterIdAndDiscussionId(Long voterId, Long discussionId);
+
+}
